Fix inefficient sql for restore statistics (#130)
Statistics are restored in 3 stages. In the first step, we update reltuples in
pg_class. In the second step, we delete statistics for a specific attribute in
pg_statistic. And at the last stage, we insert new statistics for this
attribute. At the second stage, to determine the attribute number, it is
searched in pg_attribute. For this, a subquery and the IN operator were used.
This led to an inefficient plan that contains a seq scan on pg_statistic. This,
in turn, can significantly affect the speed of statistical restore.

Fix this by replacing the IN operator with = in the attribute statistics
deletion query. We can be sure that the subquery will return no more than one
row, since pg_attribute has a unique index on the attrelid and attname
attributes.

Since this fix will only affect new backups, a fix for gprestore has also been
added. For backups created by problematic versions of gpbackup, nested loop is
enabled before restoring statistics. This should lead to the use of index scan
with fewer rows.