Turn off the orca optimizer in gprestore
When planning queries from gprestore, Orca trying to plan a query, but it almost
always fails due to the lack of support for planning queries which containing
system tables. At the same time, the start of Orca planning takes a considerable
amount of time. This is especially noticeable when restoring statistics.

This patch disables orca in gprestore sessions.