Support subset restore with on-error-continue mode (#131)
Add possibility to use on-error-continue with filtering.

Add the discardData method to the IRestoreReader interface. When an error
happens during COPY FROM execution, discardData is called to read the rest of
the table data from plugin's pipe. So the transition to the next table occurs
and the offset is positioned automatically.

In the case of discard error we don't stop the helper, because other helpers can
restore data on other segments. Another reason - hanging on COPY happens when
one helper stops before restoring a next table. Add the discardErr boolean field
to RestoreReader. This field is used in discardData and copyData to avoid
bufReader reading when an error has happened in discardData before.

When skip file is found, call discardData to move on the next table data.

Data written in writeTestTOC has now started to be used. Fix this data.

Ticket: ADBDEV-7890