Revert "Fix inefficient sql for restore statistics (#130)"
This patch fixed inefficient sql when restoring statistics by replacing the IN
operator with =. To fix existing backups, the nested loop was enabled when
restoring statistics. This was done only for a certain range of patchsets. But
it turned out that the binaries that are supplied to customers do not have a
patchset number, which is why this optimization is not activated. Therefore,
this patch has been reworked in the next commits.

This reverts commit bbbd8013a1de8cf1a0efe7e13dd4aa1aced1411d.